% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateFY.R
\name{calculateFY}
\alias{calculateFY}
\title{Calculate any fiscal year}
\usage{
calculateFY(date = Sys.Date(), date.format = "\%m/\%d/\%Y", ytd = FALSE)
}
\arguments{
\item{x}{A date object}
}
\value{
The fiscal year of the date object
}
\description{
calculate fiscal year of a date based on 7/1 - 6/30 fiscal calendar
}
