% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currentFY.R
\name{getFY}
\alias{getFY}
\title{Calculate current fiscal year}
\usage{
getFY()
}
\arguments{
\item{x}{No input}
}
\value{
The current fiscal year
}
\description{
Calculate current fiscal year based on 7/1 - 6/30 fiscal calendar
}
